refactor: model 패키지를 domain 패키지로 이름 변경
