+++ src/main/java/racingcar/controller/RaceController.java
@@ -1,8 +1,8 @@
 package racingcar.controller;
 
-import racingcar.model.car.Cars;
-import racingcar.model.car.Names;
-import racingcar.model.trycount.TryCount;
+import racingcar.domain.car.Cars;
+import racingcar.domain.car.Names;
+import racingcar.domain.trycount.TryCount;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
