+++ src/main/java/racingcar/domain/car/Cars.java
@@ -1,4 +1,4 @@
-package racingcar.model.car;
+package racingcar.domain.car;
 
 import racingcar.util.RandomGenerator;
