+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar.model.car;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
